functions/mlops_usecase/d_predictions_endpoint/app/requirements.txt
cachetools==5.3.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
db-dtypes==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-bigquery==3.21.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-firestore==2.16.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-pubsub==2.21.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.16.0 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.63.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.13.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.62.2 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.62.2 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
joblib==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==16.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
scikit-learn==1.4.2 ; python_version >= "3.11" and python_version < "4.0"
scipy==1.13.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
threadpoolctl==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
